+++ OOP_Lab_5/src/Core/Algorithms/IFindDeterminant.cs
@@ -8,6 +8,6 @@ namespace OOP_Lab_5.Core.Algorithms
 {
     public interface IFindDeterminant
     {
-        int Execute(Matrix matrix);
+        long Execute(Matrix matrix);
     }
 }
