+++ OOP_Lab_5/src/Core/Algorithms/MathNet/LibraryFindDeterminant.cs
@@ -4,12 +4,12 @@ namespace OOP_Lab_5.Core.Algorithms
 {
     public class LibraryFindDeterminant : IFindDeterminant
     {
-        public int Execute(Matrix matrix)
+        public long Execute(Matrix matrix)
         {
             var matrixBuilder = Matrix<double>.Build;
             var libMatrix = matrixBuilder.
                 Dense(matrix.Count, matrix.Count, (i, j) => matrix[i, j]);
-            return (int)libMatrix.Determinant();
+            return (long)libMatrix.Determinant();
         }
     }
 }
